pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Svet Ivanov <svetoslav.bc@gmail.com>"]
packages = [
	{ include = "page_analyzer"},
	]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.2"
flake8 = "^6.0.0"
psycopg2-binary = "^2.9.5"
bs4 = "^0.0.1"
validators = "^0.20.0"
python-dotenv = "^1.0.0"
pytest-mock = "^3.10.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
